refactor: restructure application setup by introducing factory pattern and modular configuration
- Moved application creation logic to a new factory function in app/factory.py for better organization.
- Separated configuration concerns into app/config/app_config.py, including CORS, Swagger, and database setup.
- Updated main.py to utilize the new factory function and improved debug mode configuration.
